+++ src/main/java/com/piper/valley/models/repository/ProductRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 	Optional<Product> findByName(String name);
-	Optional<Product> findByAveragePriceBetween(Double start, Double End);
+	Optional<Product>   findByAveragePriceBetween(Float start, Float End);
 	List<Product> findAll();
 }
