+++ src/main/java/com/piper/valley/models/service/ProductService.java
@@ -9,9 +9,9 @@ import java.util.Optional;
 
 public interface ProductService {
 
-	Optional<Product>getProductById(long id);
+	Optional<Product>getProductById(Integer id);
 	Optional<Product>getProductByName(String name);
-	Optional<Product>getPriceBetween(double start,double end);
+	Optional<Product>getPriceBetween(Float start,Float end);
 	Collection<Product>getAllProducts();
 	Product addProduct(AddProductForm productForm);
